+++ react-client/src/Pages/About.js
@@ -14,6 +14,17 @@ function About() {
             finalized, you'll see your shifts on the My Shifts page.
             </p> 
           </div>
+
+          <h3 className="text-center">Manager Home Page</h3>
+      <h6 className="text-center pl-5 pr-5">Welcome to your company's convenient home shift scheduler!
+        <br />
+        <br />
+        This app makes scheduling easier by giving your employees the freedom to set their availability from home.
+        We have a handy scheduling tool to make building timesheets a snap!
+        <br />
+        <br />
+        Check out the schedules tab and start building your next weekly schedule!
+      </h6>
     </>
   );
 }
