+++ react-client/src/Pages/AvailabilityPage.js
