+++ react-client/src/Pages/EmployeeHome.js
@@ -6,7 +6,7 @@ function EmployeeHome() {
       <div className="home_image">
         <h3 className="strong text-center mb-0">Welcome to <span className="blue">Schedulo!</span></h3>
         <h6 className="lead text-center mt-0">A comfortable solution to setting your work schedule around your life.</h6>
-        <div class="well width-300  align-center center centered-button">
+        <div className="well width-300  align-center center centered-button">
           
         <Link className="btn btn-primary btn-block little-opacity" to="/employee/availability">Build Your Schedule</Link>
         <Link className="btn btn-link btn-block" to="/shifts">View your Shifts.</Link>
