+++ react-client/src/Pages/SchedulesPage.js
@@ -1,0 +1,19 @@
+import Navbar from "../Components/Navbar";
+import SchedulesList from "../Components/SchedulesList";
+
+function ViewSchedules() {
+    return (
+        <>
+            <div className="flex-container">
+                <div className="schedule_image">
+                </div>
+                <div className="flex-child-list special">
+                        <h3 className="mb-0 mt-0 blue">Schedules</h3>
+                        <SchedulesList />
+                </div>
+            </div>
+        </>
+    );
+}
+
+export default ViewSchedules;
