--- react-client/src/Pages/ViewSchedules.js
@@ -1,13 +1,0 @@
-import Navbar from "../Components/Navbar";
-import SchedulesList from "../Components/SchedulesList";
-
-function ViewSchedules() {
-    return (
-        <>
-            <h3 className="text-center">Schedules</h3>
-            <SchedulesList />
-        </>
-    );
-}
-
-export default ViewSchedules;
